feat: remove workers and refactor context
