+++ package.json
@@ -29,7 +29,6 @@
     "react-hook-form": "^7.53.0",
     "react-imask": "^7.6.1",
     "react-router-dom": "^6.26.2",
-    "xlsx": "^0.18.5",
     "yup": "^1.4.0"
   },
   "devDependencies": {
