+++ src/components/PrivateRoute.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { Navigate } from "react-router-dom";
-import { useAuth } from "src/context/auth";
+import { useAuth } from "src/context/hooks";
 
 interface PrivateRouteProps {
   children: React.ReactElement;
