--- src/context/auth/AuthContext.ts
@@ -1,7 +1,0 @@
-import { createContext } from 'react';
-
-import { AuthContextType } from './AuthContext.types';
-
-const AuthContext = createContext<AuthContextType | undefined>(undefined);
-
-export default AuthContext;
