--- src/context/auth/index.ts
@@ -1,3 +1,0 @@
-export { default as AuthProvider } from './AuthProvider';
-export * from './AuthContext.types';
-export * from './AuthContext.hooks';
