+++ src/context/index.ts
@@ -1,0 +1,1 @@
+export { AuthProvider } from "./AuthContext";
