+++ src/layouts/BaseLayout/components/Drawer.tsx
@@ -23,7 +23,7 @@ import {
 } from "@mui/material";
 import { NavLink, useLocation } from "react-router-dom";
 import { VERSION } from "src/consts";
-import { useAuth } from "src/context/auth";
+import { useAuth } from "src/context/hooks";
 import { useCube } from "src/context/cube";
 
 function getRoutes(isAdmin: boolean) {
