+++ src/layouts/ClientLayout.tsx
@@ -2,7 +2,7 @@ import { useCallback } from "react";
 
 import { useNavigate } from "react-router-dom";
 import { PrivateRoute } from "src/components";
-import { useAuth } from "src/context/auth";
+import { useAuth } from "src/context/hooks";
 import { CubeProvider } from "src/context/cube";
 import { ThemeProvider } from "src/context/theme";
