+++ src/pages/client/Dashboard/Page.tsx
@@ -8,7 +8,7 @@ import { listAll, getMetadata, getDownloadURL, ref } from "firebase/storage";
 import * as pdfMake from "pdfmake/build/pdfmake";
 import * as pdfFonts from "pdfmake/build/vfs_fonts";
 import { CardButton, PageHeader } from "src/components";
-import { useAuth } from "src/context/auth";
+import { useAuth } from "src/context/hooks";
 import { storage } from "src/firebase";
 import { getError } from "src/utils";
