+++ src/pages/client/ImportPage/components/FileUploadStep.tsx
@@ -8,7 +8,7 @@ import { Alert, Box, Grid, Typography } from "@mui/material";
 import { STORAGE_PATH } from "@shared/consts";
 import { collection, doc } from "firebase/firestore";
 import { ref, uploadBytes } from "firebase/storage";
-import { useAuth } from "src/context/auth";
+import { useAuth } from "src/context/hooks";
 import { useCube } from "src/context/cube";
 import { firestore, storage } from "src/firebase";
 
@@ -45,8 +45,7 @@ const FileUpload = ({ handleOnFinish, fileResolution }: Props) => {
       handleOnFinish();
     } catch (error) {
       setError(`Error al subir el archivo: ${error}`);
-      debugger;
-      // await cube.removeCube(); // TODO: pass which cube
+      await cube.removeCube();
     } finally {
       setLoading(false);
     }
