+++ src/pages/client/ImportPage/components/index.ts
@@ -4,4 +4,3 @@ export { default as ParamsDataStep } from "./ParamsDataStep";
 export { default as DriversStep } from "./DriversStep";
 export { default as ImportedDataPage } from "./ImportedDataPage";
 export { default as ImportDataPage } from "./ImportDataPage";
-export { default as SelectSheetStep } from "./SelectSheetStep";
