+++ src/pages/client/Scorecard/hooks/useScorecard.ts
@@ -5,7 +5,7 @@ import { IScorecardData } from "@shared/models";
 import { ICallableRequest, ICallableResponse } from "@shared/models/functions";
 import { doc, getDoc, setDoc } from "firebase/firestore";
 import { httpsCallable } from "firebase/functions";
-import { useAuth } from "src/context/auth";
+import { useAuth } from "src/context/hooks";
 import { firestore, functions } from "src/firebase";
 import { getError } from "src/utils";
