+++ src/utils/file.ts
@@ -1,204 +1,4 @@
-import { EColumnType } from "@shared/enums";
-import {
-  IBaseData,
-  IDriver,
-  IParamsData,
-  IScorecardData,
-} from "@shared/models";
-import { getColumnIndex } from "@shared/utils";
-import { WorkBook } from "xlsx";
-
-export async function getWorkbookFromFileAsync(file: Blob): Promise<WorkBook> {
-  return new Promise((resolve, reject) => {
-    const worker = new Worker(
-      new URL("./workers/fileReaderWorker.js", import.meta.url)
-    );
-
-    worker.postMessage(file);
-    worker.onmessage = (event) => {
-      const { status, data, message } = event.data;
-      if (status === "success") {
-        resolve(data);
-      } else {
-        reject(new Error(message));
-      }
-      worker.terminate();
-    };
-
-    worker.onerror = (error) => {
-      reject(new Error("Error in worker: " + error.message));
-      worker.terminate();
-    };
-  });
-}
-
-export async function getJsonDataFromWorkbookAsync(
-  workbook: WorkBook,
-  sheet: string
-): Promise<unknown[]> {
-  return new Promise((resolve, reject) => {
-    const worker = new Worker(
-      new URL("./workers/sheetReaderWorker.js", import.meta.url)
-    );
-
-    worker.postMessage({ workbook, sheet });
-    worker.onmessage = (event) => {
-      const { status, data, message } = event.data;
-      if (status === "success") {
-        resolve(data);
-      } else {
-        reject(new Error(message));
-      }
-      worker.terminate();
-    };
-
-    worker.onerror = (error) => {
-      reject(new Error("Error in worker: " + error.message));
-      worker.terminate();
-    };
-  });
-}
-
-export function processJsonData(jsonData?: unknown[]): Promise<{
-  rows: { [key: string]: string | number }[];
-  columns: string[];
-}> {
-  if (!jsonData) throw new Error("No data found in the file.");
-
-  return new Promise((resolve, reject) => {
-    const worker = new Worker(
-      new URL("./workers/getColsAndRowsWorker.js", import.meta.url)
-    );
-
-    worker.onmessage = (event) => {
-      const { rows, columns, progress } = event.data;
-
-      if (progress || progress == 0) {
-        console.log(`Progress: ${progress}%`);
-      } else {
-        resolve({ rows, columns });
-      }
-    };
-
-    worker.onerror = (error) => {
-      reject(new Error("Error in worker: " + error.message));
-    };
-
-    worker.postMessage(jsonData);
-  });
-}
-
-export function getGeneralDataAsync(rows?: any[]): Promise<{
-  categories: string[];
-  sumSales: number;
-  sumCostSales: number;
-}> {
-  return new Promise((resolve, reject) => {
-    const worker = new Worker(
-      new URL("./workers/calculateGeneralDataWorker.js", import.meta.url)
-    );
-
-    worker.onmessage = (event) => {
-      const { categories, sumSales, sumCostSales, progress, error } =
-        event.data;
-
-      if (error) {
-        reject(new Error(error));
-      } else if (progress || progress == 0) {
-        console.log(`Progress: ${progress}%`);
-      } else {
-        resolve({ categories, sumSales, sumCostSales });
-      }
-    };
-
-    worker.onerror = (error) => {
-      reject(new Error("Error in worker: " + error.message));
-    };
-
-    worker.postMessage({
-      rows,
-      categoryIndex: getColumnIndex(EColumnType.CATEGORY),
-      salesIndex: getColumnIndex(EColumnType.SALES),
-      salesCostIndex: getColumnIndex(EColumnType.COST_SALES),
-    });
-  });
-}
-
-export function getCategoriesDataRowsAsync(
-  rows: any[],
-  drivers: IDriver[]
-): Promise<IBaseData["categoriesData"]["rows"]> {
-  return new Promise((resolve, reject) => {
-    const worker = new Worker(
-      new URL("./workers/calculateCategoriesDataWorker.js", import.meta.url)
-    );
-
-    worker.onmessage = (event) => {
-      const { data, progress, error } = event.data;
-
-      if (error) {
-        reject(new Error(error));
-      } else if (progress || progress == 0) {
-        console.log(`Progress: ${progress}%`);
-      } else {
-        resolve(data as IBaseData["categoriesData"]["rows"]);
-      }
-    };
-
-    worker.onerror = (error) => {
-      reject(new Error("Error in worker: " + error.message));
-    };
-
-    worker.postMessage({
-      rows,
-      category: {
-        index: getColumnIndex(EColumnType.CATEGORY)!,
-      },
-      drivers,
-    });
-  });
-}
-
-export function getCategoriesDataTotals(
-  categoriesDataRows: IBaseData["categoriesData"]["rows"],
-  drivers: IDriver[]
-): IBaseData["categoriesData"]["totals"] {
-  return {
-    category: "Total",
-    ...drivers
-      .filter((x) => -1 !== x.columnIndexReference)
-      .reduce(
-        (acc, driver) => {
-          acc[driver.key] = categoriesDataRows.reduce(
-            (acc, row) => acc + (row[driver.key] as number),
-            0
-          );
-          return acc;
-        },
-        {} as Omit<IBaseData["categoriesData"]["totals"], "category">
-      ),
-  } as IBaseData["categoriesData"]["totals"];
-}
-
-export function getDriversDataRows(
-  drivers: IDriver[],
-  rows: IBaseData["categoriesData"]["rows"],
-  totals: IBaseData["categoriesData"]["totals"]
-): IBaseData["driversData"]["rows"] {
-  return drivers.map((driver) => {
-    const _row = {
-      driver: driver.label,
-    } as IBaseData["driversData"]["rows"][number];
-
-    rows.forEach((row) => {
-      _row[row.category] = Number(totals[driver.key])
-        ? Number(row[driver.key]) / Number(totals[driver.key])
-        : 0;
-    });
-
-    return _row;
-  });
-}
+import { IBaseData, IParamsData, IScorecardData } from "@shared/models";
 
 export function updateStoringScorecardDataRow(
   newRow: IScorecardData["storingCosts"]["rows"][number],
